[FJ-95] update imports for relocated Table component
