+++ frontend/src/Components/Tables/DepartmentalSearchCountTable.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // Components
-import Table from 'Components/Elements/Table';
+import Table from 'Components/Elements/Table/Table';
 
 function DepartmentalSearchCountTable({ data }) {
   const columns = React.useMemo(
