+++ frontend/src/Components/Tables/DepartmentalSearchRateTable.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // Components
-import Table from 'Components/Elements/Table';
+import Table from 'Components/Elements/Table/Table';
 
 function DepartmentalSearchRateTable({ stops, searches }) {
   const columns = React.useMemo(
