+++ frontend/src/Components/Tables/StopsByEthnicCompositionTable.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // Components
-import Table from 'Components/Elements/Table';
+import Table from 'Components/Elements/Table/Table';
 
 function StopsByEthnicCompositionTable({ data }) {
   const columns = React.useMemo(
